console: Move application actions to new file move action creators to actions/applications.js
