+++ pkg/webui/console/containers/application-events/index.js
@@ -21,7 +21,7 @@ import EventsSubscription from '../../containers/events-subscription'
 
 import {
   clearApplicationEventsStream,
-} from '../../store/actions/application'
+} from '../../store/actions/applications'
 
 import {
   selectApplicationEvents,
