+++ pkg/webui/console/store/reducers/index.js
@@ -13,8 +13,10 @@
 // limitations under the License.
 
 import { combineReducers } from 'redux'
-import { SHARED_NAME as APPLICATION_SHARED_NAME } from '../actions/application'
-import { SHARED_NAME as APPLICATIONS_SHARED_NAME } from '../actions/applications'
+import {
+  SHARED_NAME_SINGLE as APPLICATION_SHARED_NAME,
+  SHARED_NAME as APPLICATIONS_SHARED_NAME,
+} from '../actions/applications'
 import { SHARED_NAME as GATEWAY_SHARED_NAME } from '../actions/gateway'
 import { SHARED_NAME as GATEWAYS_SHARED_NAME } from '../actions/gateways'
 import { SHARED_NAME as DEVICE_SHARED_NAME } from '../actions/device'
