+++ pkg/webui/console/views/application-api-key-edit/index.js
@@ -26,8 +26,10 @@ import Message from '../../../lib/components/message'
 import IntlHelmet from '../../../lib/components/intl-helmet'
 import { ApiKeyEditForm } from '../../../components/api-key-form'
 
-import { getApplicationApiKey } from '../../store/actions/application'
-import { getApplicationsRightsList } from '../../store/actions/applications'
+import {
+  getApplicationApiKey,
+  getApplicationsRightsList,
+} from '../../store/actions/applications'
 import {
   selectSelectedApplicationId,
   selectApplicationRights,
